refactor - move factory under provider package
