+++ cmd/egress/cmd.go
@@ -7,7 +7,7 @@ import (
 
 	ocmlog "github.com/openshift-online/ocm-sdk-go/logging"
 	"github.com/openshift/osd-network-verifier/pkg/cloudclient"
-	"github.com/openshift/osd-network-verifier/pkg/parameters"
+	"github.com/openshift/osd-network-verifier/pkg/cloudclient/aws"
 	"github.com/spf13/cobra"
 )
 
@@ -40,8 +40,8 @@ are set correctly before execution.
 	//client args - all these have defaults
 	validateEgressCmd.Flags().StringVar(&config.CloudImageID, "image-id", "", "(optional) cloud image for the compute instance")
 	validateEgressCmd.Flags().StringVar(&config.InstanceType, "instance-type", "t3.micro", "(optional) compute instance type")
-	validateEgressCmd.Flags().StringVar(&config.Region, "region", config.Region, fmt.Sprintf("(optional) compute instance region. If absent, environment var %[1]v will be used, if set", cloudclient.RegionEnvVarStrAWS, cloudclient.RegionDefaultAWS))
-	validateEgressCmd.Flags().StringToStringVar(&config.CloudTags, "cloud-tags", cloudclient.DefaultTagsAWS, "(optional) comma-seperated list of tags to assign to cloud resources e.g. --cloud-tags key1=value1,key2=value2")
+	validateEgressCmd.Flags().StringVar(&config.Region, "region", config.Region, fmt.Sprintf("(optional) compute instance region. If absent, environment var %[1]v will be used, if set", aws.RegionEnvVarStrAWS, aws.RegionDefaultAWS))
+	validateEgressCmd.Flags().StringToStringVar(&config.CloudTags, "cloud-tags", aws.DefaultTagsAWS, "(optional) comma-seperated list of tags to assign to cloud resources e.g. --cloud-tags key1=value1,key2=value2")
 	validateEgressCmd.Flags().BoolVar(&config.Debug, "debug", false, "(optional) if true, enable additional debug-level logging")
 	validateEgressCmd.Flags().DurationVar(&config.Timeout, "timeout", cloudclient.DefaultTime, "(optional) timeout for individual egress verification requests")
 	validateEgressCmd.Flags().StringVar(&config.KmsKeyID, "kms-key-id", "", "(optional) ID of KMS key used to encrypt root volumes of compute instances. Defaults to cloud account default key")
@@ -71,7 +71,7 @@ func rune(cmd *cobra.Command, args []string) error {
 	}
 
 	//Downstream must pass in data in the form defined in parameters.ValidateEgress
-	out := client.ValidateEgress(parameters.ValidateEgress{
+	out := client.ValidateEgress(cloudclient.ValidateEgress{
 		VpcSubnetID: inputStruct.vpcSubnetId, // required downstream argument
 	})
